backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0

# Database
aiosqlite>=0.20.0

# AI Providers
anthropic>=0.40.0
google-genai>=1.0.0

# Image Processing
pillow>=10.0.0

# Export (Plan 005)
python-pptx>=1.0.0
playwright>=1.49.0

# File processing
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.0

# Utilities
structlog>=24.4.0
python-multipart>=0.0.9

